Refactor repository interfaces: rename and relocate contracts
