+++ Library.Domain/Library.Domain.csproj
@@ -6,4 +6,8 @@
 		<Nullable>enable</Nullable>
 	</PropertyGroup>
 
+	<ItemGroup>
+	  <ProjectReference Include="..\Library.Application.Contracts\Library.Application.Contracts.csproj" />
+	</ItemGroup>
+
 </Project>
