+++ Library.Infrastructure.MongoEf/Repositories/IssueRepository.cs
@@ -1,5 +1,5 @@
 ﻿using Library.Domain.Models;
-using Library.Infrastructure.MongoEf.Contracts;
+using Library.Domain.RepositoryInterfaces;
 using Library.Infrastructure.MongoEf.Database;
 using Microsoft.EntityFrameworkCore;
